Add initial custom styling
New classes allow for setting styles of elements in the
calendar: day, holiday and moon phase styles. This customizes
placement of text for all three. For the moon phases, a specific
font can be selected, which helps pick the font in Windows 11
(Cambria Math Regular) that features all the glyphs usable for
displaying the moon phases.