Introduce calendar style for reduced row count
The new class describes the calendar styles. The first property is to
select between a full (always 6 weeks displayed each week in its own
row) calendar or a reduced (always 5 weeks displayed, any 6th week days
overlap with the coincidental week 5 days).

For the days that have week 5 and week 6 overlaps, the custom playement
styles place the days in the apropriate corner of the cell.

Still TODO for the overlapping days:
* Placement of holidays.
* Placement of moon phases.
* A diagonal line separating the days.