Extend styling options for holidays
The "FLOP" (First Line Offset) for holidays text is now optional and
can be enabled or disabled in the style.

The distances from borders are now also customizable. The code does
change the behaviour for the default style in that it sets the 5% row
height or width for left, right and top distance as well, but I figured
this is OK since the classic style uses bottom alignment.